exif-tool/requirements.txt
Flask==3.1.2
Pillow==12.1.0
geopy==2.4.1
gunicorn==21.2.0
